requirements.txt
altgraph==0.17.2
blinker==1.7.0
click==8.1.7
Flask==3.0.0
future==0.18.2
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.15.2
MarkupSafe==2.1.3
multiline==1.0.2
numpy==1.26.2
pandas==2.1.4
pika==1.3.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==58.0.4
six==1.15.0
tzdata==2023.3
Werkzeug==3.0.1
wheel==0.37.0
zipp==3.17.0
